+++ ola_core/src/sequencer/seal_criteria/conditional_sealer.rs
@@ -39,7 +39,7 @@ impl ConditionalSealer {
 
     pub(super) fn should_seal_l1_batch(
         &self,
-        l1_batch_number: u32,
+        l1_batch_number: u64,
         block_open_timestamp_ms: u128,
         tx_count: usize,
         block_data: &SealData,
